provides frameworks for host and guest clusters
