+++ pkg/framework/wait.go
@@ -1,10 +1,6 @@
 package framework
 
 import (
-	"log"
-	"os"
-	"os/exec"
-	"strings"
 	"time"
 
 	"github.com/cenkalti/backoff"
@@ -14,17 +10,6 @@ import (
 const (
 	defaultTimeout = 600
 )
-
-func runCmd(cmdStr string) error {
-	log.Printf("Running command %q\n", cmdStr)
-	cmdEnv := os.ExpandEnv(cmdStr)
-	fields := strings.Fields(cmdEnv)
-	cmd := exec.Command(fields[0], fields[1:]...)
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stdout
-
-	return cmd.Run()
-}
 
 func waitFor(f func() error) error {
 	return baseWait(backoff.NewExponentialBackOff(), f)
